requirements.txt
Flask==2.3.3
tensorflow==2.18.0
numpy==1.26.4
Pillow==10.3.0